+++ fs/xfs/xfs_dir.c
@@ -139,17 +139,17 @@ xfs_dir_mount(xfs_mount_t *mp)
 {
 	uint shortcount, leafcount, count;
 
-	shortcount = (mp->m_attroffset - sizeof(xfs_dir_sf_hdr_t)) /
-		     sizeof(xfs_dir_sf_entry_t);
-	leafcount = (XFS_LBSIZE(mp) - sizeof(xfs_dir_leaf_hdr_t)) /
-		    (sizeof(xfs_dir_leaf_entry_t) +
-		     sizeof(xfs_dir_leaf_name_t));
+	shortcount = (mp->m_attroffset - (uint)sizeof(xfs_dir_sf_hdr_t)) /
+		     (uint)sizeof(xfs_dir_sf_entry_t);
+	leafcount = (XFS_LBSIZE(mp) - (uint)sizeof(xfs_dir_leaf_hdr_t)) /
+		    ((uint)sizeof(xfs_dir_leaf_entry_t) +
+		     (uint)sizeof(xfs_dir_leaf_name_t));
 	count = shortcount > leafcount ? shortcount : leafcount;
 	mp->m_dircook_elog = xfs_da_log2_roundup(count + 1);
 	ASSERT(mp->m_dircook_elog <= mp->m_sb.sb_blocklog);
 	mp->m_da_node_ents =
-		(XFS_LBSIZE(mp) - sizeof(xfs_da_node_hdr_t)) /
-		sizeof(xfs_da_node_entry_t);
+		(XFS_LBSIZE(mp) - (uint)sizeof(xfs_da_node_hdr_t)) /
+		(uint)sizeof(xfs_da_node_entry_t);
 	mp->m_da_magicpct = (XFS_LBSIZE(mp) * 37) / 100;
 }
